Separate CommandBuffer to buffer.py
